fix(tests): move assertSecretsInputIsSecret to build-tagged file
The function was defined in main_test.go (no build tag) but only called
from files with language-specific build tags, causing the linter to flag
it as unused when no tags were active. Moving it to secrets_test.go with
matching build tags resolves the lint failure.
